Improve exception handling in branch data code
This commit introduces an enhancement to the exception handling in the BranchData.cs file. Before, exceptions were simply thrown but now, the code is implemented to show the MessageBox with the exception message. Besides, a small grammar mistake in a MessageBox message was corrected. This makes it easy to debug the issue because developers can now understand what exactly went wrong based on the exception message shown.