Update project dependencies and settings
The TargetFrameworkVersion has been updated to v4.8 to ensure compatibility with the latest .NET Framework features. Several references to third-party libraries have been upgraded to newer versions including EntityFramework, envdte and others for improved performance and security. Removed unnecessary imports and files, and reorganised some elements for better clarity and understanding of the .csproj file. The changes also include an adjustment to the .NET language version (LangVersion) to 8 (C# 8.0) allowing the use of more recent language features. All these changes are important in keeping the project up-to-date with the latest development practices.