Update package versions and target framework
Updated multiple packages to their newer versions to ensure compatibility with updated systems and to benefit from fixes, improvements or new features of the packages. Also, target framework of the packages has been updated to .NET 4.8, adjusting to the modern development patterns and requirements. This should also help with performance improvements and staying up-to-date with current technologies.